impl delete file from directory

+++ backend/src/main/java/com/malina/siteForInteriorDesigner/controller/PortfolioController.java
@@ -29,8 +29,8 @@ public class PortfolioController {
         String fileName = file.getOriginalFilename();
         System.out.println("FILENAME : " + fileName);
 
-        Path path = Paths.get(uploadDir + fileName);
-//        Files.createDirectories(path.getParent());
+        Path path = Paths.get(uploadDir + "/" + fileName);
+        Files.createDirectories(path.getParent());
         Files.write(path, file.getBytes());
 
         PortfolioEntity photo = new PortfolioEntity();
@@ -66,14 +66,24 @@ public class PortfolioController {
     }
     @DeleteMapping("/portfolio/{id}")
     public String deleteImage(@PathVariable("id") long id){
+        PortfolioEntity imageToDelete = service.getImageById(id);
+        if (imageToDelete == null) {
+            return "Error : image not found";
+        }
+
         service.deleteImage(id);
 
-        //Обновляет index после удаления фото
-//        List<PortfolioEntity> images = new ArrayList<>();
-//        images.stream()
-//                .sorted(Comparator.comparingInt(PortfolioEntity::getIndex))
-//                .forEach(entity -> entity.setIndex(images.indexOf(entity)));
-        return "Image was removed successfully...";
+        String imagePath = imageToDelete.getPath();
+        String fileName = imagePath.substring(imagePath.lastIndexOf("/") + 1);
+        Path filePath = Paths.get(uploadDir + "/" + fileName);
+
+        try {
+            Files.delete(filePath);
+        } catch (IOException e) {
+            return "Error : deleting file";
+        }
+
+        return "Image was removed successfully";
     }
